Move DNS resolution into `Io` component
